+++ numpyro/distributions/discrete.py
@@ -30,7 +30,6 @@ import warnings
 import numpy as np
 
 from jax import device_put, lax
-from jax.dtypes import canonicalize_dtype
 from jax.nn import softmax, softplus
 import jax.numpy as jnp
 import jax.random as random
@@ -354,54 +353,6 @@ def Categorical(probs=None, logits=None, validate_args=None):
         return CategoricalLogits(logits, validate_args=validate_args)
     else:
         raise ValueError('One of `probs` or `logits` must be specified.')
-
-
-class Delta(Distribution):
-    arg_constraints = {'v': constraints.real, 'log_density': constraints.real}
-    support = constraints.real
-    is_discrete = True
-
-    def __init__(self, v=0., log_density=0., event_dim=0, validate_args=None, value=None):
-        if value is not None:
-            v = value
-            warnings.warn("`value` argument has been deprecated in favor of `v` argument.",
-                          FutureWarning)
-
-        if event_dim > jnp.ndim(v):
-            raise ValueError('Expected event_dim <= v.dim(), actual {} vs {}'
-                             .format(event_dim, jnp.ndim(v)))
-        batch_dim = jnp.ndim(v) - event_dim
-        batch_shape = jnp.shape(v)[:batch_dim]
-        event_shape = jnp.shape(v)[batch_dim:]
-        self.v = lax.convert_element_type(v, canonicalize_dtype(jnp.float64))
-        # NB: following Pyro implementation, log_density should be broadcasted to batch_shape
-        self.log_density = promote_shapes(log_density, shape=batch_shape)[0]
-        super(Delta, self).__init__(batch_shape, event_shape, validate_args=validate_args)
-
-    def sample(self, key, sample_shape=()):
-        shape = sample_shape + self.batch_shape + self.event_shape
-        return jnp.broadcast_to(device_put(self.v), shape)
-
-    @validate_sample
-    def log_prob(self, value):
-        log_prob = jnp.log(value == self.v)
-        log_prob = sum_rightmost(log_prob, len(self.event_shape))
-        return log_prob + self.log_density
-
-    @property
-    def mean(self):
-        return self.v
-
-    @property
-    def variance(self):
-        return jnp.zeros(self.batch_shape + self.event_shape)
-
-    def tree_flatten(self):
-        return (self.v, self.log_density), self.event_dim
-
-    @classmethod
-    def tree_unflatten(cls, aux_data, params):
-        return cls(*params, event_dim=aux_data)
 
 
 class OrderedLogistic(CategoricalProbs):
